# internal/models/vocabularyDTO.go
package models
type VocabularyDTO struct {
	English    string    `json:"english" `
	Vietnamese string    `json:"vietnamese" `
	Tag        []string  `json:"tag" `
	Mp3        string    `json:"mp3" `
}